Add LegacyGOAuth handling
It's too painful to test this client without being able to use legacy
tokens, so the support is needed.
Also fix a couple of minor bugs.

As a bit of a breach of established patterns, the client accepts the
legacy token as a kwarg that overwrites any given authorizer. This makes
it possible to expose NexusClient without needing to expose the
authorizer as well, so it seems worth it.